Split organizer view of Pairings tab into separate route
