+++ frontend/src/routes/tournaments/[tournamentId]/organizer/rounds/+page.ts
@@ -1,5 +1,5 @@
 import type { PageLoad } from "./$types";
-import { loadPairings } from "../api_helper";
+import { loadPairings } from "../../api_helper";
 
 export const load: PageLoad = async ({ params, fetch }) => {
   const pairingsData = await loadPairings(parseInt(params.tournamentId), null, fetch);
