+++ frontend/src/routes/tournaments/[tournamentId]/organizer/rounds/AdminReportOptions.svelte
